AC/requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
PyJWT==1.7.1
UNKNOWN==0.0.0
Werkzeug==1.0.1